+++ packages/riverpod/lib/experimental/mutation.dart
@@ -1,5 +1,7 @@
 export '../src/internals.dart'
     show
+        // ignore: deprecated_member_use_from_same_package
+        MutationTransaction,
         Mutation,
         MutationState,
         MutationIdle,